:truck: rename page to scene
