+++ frontend/src/views/App.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { Route, Switch } from 'react-router-dom';
 import { ConnectedRouter } from 'connected-react-router';
 import configureStore, { history } from '../state/store';
-import LoginPage from './components/auth/LoginPage';
-import HomePage from './components/home/HomePage';
+import LoginScene from './scenes/LoginScene';
+import HomeScene from './scenes/HomeScene';
 import AuthGroupContainer from './containers/auth/AuthGroupContainer';
 
 const store = configureStore();
@@ -14,10 +14,10 @@ const App: React.FC = () => {
     <Provider store={store}>
       <ConnectedRouter history={history}>
         <Switch>
-          <Route exact path="/login" component={LoginPage} />
+          <Route exact path="/login" component={LoginScene} />
           <AuthGroupContainer>
             <Switch>
-              <Route exact path="/" component={HomePage} />
+              <Route exact path="/" component={HomeScene} />
             </Switch>
           </AuthGroupContainer>
         </Switch>
